chore: move getHomeDir to new file
